moved metric callback to separate file
